Data.Drasil.Domains

Description

Defines Domains where concepts live

Documentation

civilEng :: IdeaDict

For ideas, concepts, or terms related to Civil Engineering.

Various domains that are used in Drasil. May have an abbreviation.

materialEng :: IdeaDict

For ideas, concepts, or terms related to Material Engineering.

Various domains that are used in Drasil. May have an abbreviation.

mathematics :: IdeaDict

For ideas, concepts, or terms related to Mathematics.

Various domains that are used in Drasil. May have an abbreviation.

physics :: IdeaDict

For ideas, concepts, or terms related to Physics.

Various domains that are used in Drasil. May have an abbreviation.