Data.Drasil.Phrase

Documentation

and_ :: (NamedIdea c, NamedIdea d) => c -> d -> NP

Creates an NP by combining two NamedIdeas with the word "and" between their terms. Plural is defaulted to (phrase t1) "of" (plural t2)

and_' :: (NamedIdea c, NamedIdea d) => c -> d -> NP

Same as and_ combinator, except phrase default of second term is plural instead of phrase

andRT :: (NamedIdea c, NamedIdea d) => (c -> Sentence) -> (d -> Sentence) -> c -> d -> NP

Customizable and combinator

with :: (NamedIdea c, NamedIdea d) => c -> d -> NP

of_ :: (NamedIdea c, NamedIdea d) => c -> d -> NP

Creates a noun phrase by combining two NamedIdeas with the word "of" between their terms. Plural is defaulted to (phrase t1) "of" (plural t2)

ofN_ :: (NamedIdea c, NounPhrase d) => c -> d -> NP

of_' :: (NamedIdea c, NamedIdea d) => c -> d -> NP

Creates a noun phrase by combining two NamedIdeas with the word "of" between them. phrase is defaulted to (phrase t1) "of" (plural t2). Plural is the same.

of_'' :: (NamedIdea c, NamedIdea d) => c -> d -> NP

Same as of_, except plural default of second term is phrase

of__ :: (NamedIdea c, NamedIdea d) => c -> d -> NP

Same as of_, except phrase default of first term is plural instead of phrase

ofA :: (NamedIdea c, NamedIdea d) => c -> d -> NP

Same as of__, except combining Sentence piece is "of a"

for :: (NamedIdea c, NamedIdea d) => c -> d -> Sentence

Inserts the word "for" between the titleized versions of two terms

for' :: (NamedIdea c, Idea d) => c -> d -> Sentence

Similar to for, but uses titleized version of term 1 with the abbreviation (if it exists, phrase otherwise) for term 2

for'' :: (c -> Sentence) -> (d -> Sentence) -> c -> d -> Sentence

Similar to for, but allows one to specify the function to use on each term before inserting for. For example one could use for'' phrase plural t1 t2

the' :: NamedIdea t => t -> NP

the :: NamedIdea t => t -> NP

theCustom :: (t -> Sentence) -> t -> NP

compoundNC :: (NamedIdea a, NamedIdea b) => a -> b -> NamedChunk

Combinator for combining two NamedChunks into one. Does not preserve abbreviations

compoundNC' :: (NamedIdea a, NamedIdea b) => a -> b -> NamedChunk

compoundNC'' :: (NamedIdea a, NamedIdea b) => (NP -> Sentence) -> (NP -> Sentence) -> a -> b -> NamedChunk

compoundNCPlPh :: NamedChunk -> NamedChunk -> NamedChunk

compoundNCPlPl :: NamedChunk -> NamedChunk -> NamedChunk

compoundNC''' :: (NamedIdea a, NamedIdea b) => (NP -> Sentence) -> a -> b -> NamedChunk

compoundNCP1 :: NamedChunk -> NamedChunk -> NamedChunk