Data.Drasil.SentenceStructures

Documentation

foldlSent :: [Sentence] -> Sentence

partial function application of foldle for sentences specifically

foldlSent_ :: [Sentence] -> Sentence

foldlSent but does not end with period

foldlSentCol :: [Sentence] -> Sentence

foldlSent but ends with colon

foldlsC :: [Sentence] -> Sentence

creates a list of elements separated by commas, including the last element

foldlList :: SepType -> FoldType -> [Sentence] -> Sentence

creates a list of elements separated by a "separator", ending with "and" or "or"

foldlEnumList :: EnumType -> WrapType -> SepType -> FoldType -> [Sentence] -> Sentence

creates an list of elements with "enumerators" in "wrappers" using foldlList

tableShows :: LabelledContent -> Sentence -> Sentence

underConsidertn :: ConceptChunk -> Sentence

showingCxnBw :: NamedIdea c => c -> Sentence -> Sentence

refineChain :: NamedIdea c => [(c, Section)] -> Sentence

Create a list in the pattern of "The are refined to the ". Note: Order matters!

foldlSP :: [Sentence] -> Contents

fold sentences then turns into content

foldlSP_ :: [Sentence] -> Contents

foldlSPCol :: [Sentence] -> Contents

maybeChanged :: Sentence -> Sentence -> Sentence

maybeExpanded :: Sentence -> Sentence -> Sentence

maybeWOVerb :: Sentence -> Sentence -> Sentence

tAndDWAcc :: Concept s => s -> ItemType

helpful combinators for making Sentences for Terminologies with Definitions term (acc) - definition

tAndDWSym :: (Concept s, Quantity a) => s -> a -> ItemType

tAndDOnly :: Concept s => s -> ItemType

follows :: (Referable r, HasShortName r) => Sentence -> r -> Sentence

getTandS :: Quantity a => a -> Sentence

Used when you want to say a term followed by its symbol. ex. "...using the Force F in..."

getTDS :: (Quantity a, Concept a) => a -> Sentence

get term, definition, and symbol

eqN :: Int -> Sentence

displayConstrntsAsSet :: Quantity a => a -> [String] -> Sentence

chgsStart :: (HasShortName x, Referable x) => x -> Sentence -> Sentence

data EnumType

Constructors

Numb

Upper

Lower

data WrapType

Constructors

Parens

Period

data SepType

Constructors

Comma

SemiCol

data FoldType

Constructors

List

Options