Data.Drasil.Utils

Documentation

addPercent :: Show a => a -> Sentence

outputs sentence with % attached to it

bulletFlat :: [Sentence] -> ListType

bulletFlat applies Bullet and Flat to a list.

bulletNested :: [Sentence] -> [ListType] -> ListType

bulletNested applies Bullets and headers to a Nested ListType. t - Headers of the Nested lists. l - Lists of ListType.

enumBullet :: Reference -> [Sentence] -> LabelledContent

enumBullet apply Enumeration, Bullet and Flat to a list

enumBulletU :: [Sentence] -> Contents

enumSimple :: Reference -> Integer -> Sentence -> [Sentence] -> LabelledContent

enumSimple enumerates a list and applies simple and enumeration to it s - start index for the enumeration t - title of the list l - list to be enumerated

enumSimpleU :: Integer -> Sentence -> [Sentence] -> Contents

eqUnR :: Expr -> Reference -> LabelledContent

eqUnR' :: Expr -> Contents

fmtPhys :: (Constrained c, Quantity c) => c -> Sentence

formats physical constraints

fmtSfwr :: (Constrained c, Quantity c) => c -> Sentence

formats software constraints

fmtU :: MayHaveUnit a => Sentence -> a -> Sentence

takes a amount and adds a unit to it n - sentenc representing an amount u - unit we want to attach to amount

getRVal :: (HasUID c, HasReasVal c) => c -> Expr

gets a reasonable or typical value from a Constrained chunk

itemRefToSent :: String -> Sentence -> Sentence

makes sentences from an item and its reference a - String title of reference b - Sentence containing the full reference

makeListRef :: [a] -> Section -> [Sentence]

makeListRef takes a list and a reference and generates references to match the length of the list l - list whos length is to be matched

makeTMatrix :: Eq a => [Sentence] -> [[a]] -> [a] -> [[Sentence]]

makes a traceability matrix from list of column rows and list of rows

mkEnumAbbrevList :: Integer -> Sentence -> [Sentence] -> [(Sentence, ItemType)]

zip helper function enumerates abbreviation and zips it with list of itemtype s - the number from which the enumeration should start from t - the title of the list l - the list to be enumerated

mkTableFromColumns :: [(Sentence, [Sentence])] -> ([Sentence], [[Sentence]])

Helper for making a table from a columns

noRefs :: [ItemType] -> [(ItemType, Maybe String)]

noRefs converts lists of simple ItemTypes into a lists which may be used in Contents but not directly referable.

noRefsLT :: [(Sentence, ItemType)] -> [ListTuple]

noRefsLT converts lists of tuples containing a title and ItemType into a ListTuple which can be used with Contents but not directly referable.

typUncr :: HasUncertainty c => c -> Sentence

extracts the typical uncertainty to be displayed from something that has an uncertainty

unwrap :: Maybe UnitDefn -> Sentence

get a unit symbol if there is one

weave :: [[a]] -> [a]

interweaves two lists together [[a,b,c],[d,e,f]] -> [a,d,b,e,c,f]

zipFTable' :: Eq a => [a] -> [a] -> [Sentence]

traceability matrices row from a list of rows and a list of columns

zipSentList :: [[Sentence]] -> [Sentence] -> [[Sentence]] -> [[Sentence]]

appends a sentence to the front of a list of list of sentences